requirements.txt
pytesseract
pillow
PyPDF2